cleanup downloader integraion
