+++ go.mod
@@ -1,5 +1,5 @@
 module github.com/bobiverse/docxplate
 
-go 1.22
+go 1.23.1
 
-require golang.org/x/text v0.16.0
+require golang.org/x/text v0.18.0
